项目文档/3-固件工具/05-雷达配置参数研究/requirements.txt
pyserial
psutil
CairoSVG
Pillow
